refactor: 뉴스 페이지 제거 및 응급처치 백과사전 페이지로 교체
