+++ client/public/locales/en/common.json
@@ -14,7 +14,7 @@
   "navbar": {
     "home": "Home",
     "about": "About",
-    "news": "News",
+    "news": "Encyclopedia",
     "translate": "AI Translation",
     "first_aid": "AI First Aid",
     "hospital_pharmacy_nearby": "Hospital & Pharmacy Nearby"
