+++ client/public/locales/ko/common.json
@@ -14,7 +14,7 @@
   "navbar": {
     "home": "홈",
     "about": "소개",
-    "news": "뉴스",
+    "news": "응급처치 백과사전",
     "translate": "AI 번역",
     "first_aid": "AI 응급처치",
     "hospital_pharmacy_nearby": "병원 & 약국 찾기"
