--- client/src/features/news/api/newsApi.ts
@@ -1,36 +1,0 @@
-import { httpServer } from '@/src/shared/utils/httpServer';
-import { NewsQueryParams, NewsResponse } from '../types';
-
-// ISR 전용 함수
-export async function fetchMedicalNewsSSR({
-  page = 1,
-  pageSize = 10,
-  country = 'us',
-}: NewsQueryParams = {}): Promise<NewsResponse> {
-  try {
-    const response = await httpServer.get<{
-      message: string;
-      data: {
-        articles: NewsResponse['articles'];
-        totalResults: number;
-        page: number;
-        pageSize: number;
-      };
-    }>(`/news/headlines?page=${page}&pageSize=${pageSize}&country=${country}`, {
-      next: {
-        tags: ['medical-news', `news-page-${page}`],
-        revalidate: 3600,
-      },
-    });
-
-    return response.data;
-  } catch (error) {
-    console.warn('Failed to fetch medical news:', error);
-    return {
-      articles: [],
-      totalResults: 0,
-      page,
-      pageSize,
-    };
-  }
-}
