--- client/src/features/news/index.ts
@@ -1,13 +1,0 @@
-export type { NewsArticle, NewsQueryParams, NewsResponse } from './types';
-
-export { fetchMedicalNewsSSR } from './api/newsApi';
-
-export {
-  LoadingSkeleton,
-  NewsCard,
-  NewsFooter,
-  NewsGrid,
-  NewsHeader,
-  NewsPageClient,
-  Pagination,
-} from './ui';
