--- client/src/features/news/ui/NewsHeader.tsx
@@ -1,11 +1,0 @@
-export function NewsHeader() {
-  return (
-    <div className='mb-12 text-center'>
-      <h1 className='mb-4 text-4xl font-bold text-gray-900'>Medical News</h1>
-      <p className='mx-auto max-w-2xl text-lg text-gray-600'>
-        Stay updated with the latest medical news, health research, and healthcare developments from
-        around the world.
-      </p>
-    </div>
-  );
-}
