--- client/src/features/news/ui/Pagination.tsx
@@ -1,79 +1,0 @@
-interface PaginationProps {
-  currentPage: number;
-  totalPages: number;
-  onPageChange: (page: number) => void;
-}
-
-export function Pagination({ currentPage, totalPages, onPageChange }: PaginationProps) {
-  const getVisiblePages = () => {
-    const delta = 2;
-    const range = [];
-    const rangeWithDots = [];
-
-    for (
-      let i = Math.max(2, currentPage - delta);
-      i <= Math.min(totalPages - 1, currentPage + delta);
-      i++
-    ) {
-      range.push(i);
-    }
-
-    if (currentPage - delta > 2) {
-      rangeWithDots.push(1, '...');
-    } else {
-      rangeWithDots.push(1);
-    }
-
-    rangeWithDots.push(...range);
-
-    if (currentPage + delta < totalPages - 1) {
-      rangeWithDots.push('...', totalPages);
-    } else {
-      if (totalPages > 1) rangeWithDots.push(totalPages);
-    }
-
-    return rangeWithDots;
-  };
-
-  const visiblePages = getVisiblePages();
-
-  return (
-    <div className='mt-8 flex items-center justify-center space-x-2'>
-      <button
-        onClick={() => onPageChange(currentPage - 1)}
-        disabled={currentPage === 1}
-        className='rounded-md border border-gray-300 bg-white px-3 py-2 text-sm font-medium text-gray-500 hover:bg-gray-50 disabled:cursor-not-allowed disabled:opacity-50'
-      >
-        Previous
-      </button>
-
-      {visiblePages.map((page, index) =>
-        typeof page === 'number' ? (
-          <button
-            key={page}
-            onClick={() => onPageChange(page)}
-            className={`rounded-md px-3 py-2 text-sm font-medium ${
-              currentPage === page
-                ? 'border border-blue-600 bg-blue-600 text-white'
-                : 'border border-gray-300 bg-white text-gray-500 hover:bg-gray-50'
-            }`}
-          >
-            {page}
-          </button>
-        ) : (
-          <span key={`dots-${index}`} className='px-2 py-2 text-gray-400'>
-            {page}
-          </span>
-        ),
-      )}
-
-      <button
-        onClick={() => onPageChange(currentPage + 1)}
-        disabled={currentPage === totalPages}
-        className='rounded-md border border-gray-300 bg-white px-3 py-2 text-sm font-medium text-gray-500 hover:bg-gray-50 disabled:cursor-not-allowed disabled:opacity-50'
-      >
-        Next
-      </button>
-    </div>
-  );
-}
