--- client/src/features/news/ui/index.ts
@@ -1,8 +1,0 @@
-export { LoadingSkeleton } from './LoadingSkeleton';
-export { NewsCard } from './NewsCard';
-export { NewsFooter } from './NewsFooter';
-export { NewsGrid } from './NewsGrid';
-export { NewsHeader } from './NewsHeader';
-export { NewsPageClient } from './NewsPageClient';
-export { Pagination } from './Pagination';
-export { RevalidateButton } from './RevalidateButton';
