+++ client/src/widgets/bottomNavigateBar/ui/ConditionalBottomNavigateBar.tsx
@@ -8,7 +8,6 @@ export default function ConditionalBottomNavigateBar() {
 
   const shouldHide =
     pathname.startsWith('/about') ||
-    pathname.startsWith('/news') ||
     pathname.startsWith('/triage') ||
     pathname === '/login' ||
     pathname === '/signup' ||
